Adding Globals Styles, and Font-Face
Now in the globals css files we have the Audiowide font
face that is accessible via a variable --chap-font.
Among the font variable I have added others for sizing
and colors. I have also changed the font-size of the
html ot be 62.5% meaning the base unit of a rem is now
10px making it easier to know by exactly how much you
are scaling.

I've introduced the notion of variables in this way so
that way we can be consistent with out colors and make
it easier to be responsive with out sizing. Instead of
adding @media css styling to each component we can just
do so for the html. Changing the font size from 62.5%
to 50% when we get down to a certain size and everything
will resize accordingly.